refactor after release of assessment result

+++ src/main/java/sg/edu/nus/iss/app/assessment/model/Account.java
@@ -1,5 +1,7 @@
 package sg.edu.nus.iss.app.assessment.model;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,5 +13,5 @@ public class Account {
 
     private String accountId;
     private String name;
-    private float balance;
+    private BigDecimal balance;
 }
